requirements.txt
# Core Trading Frameworks
backtrader==1.9.78.123
ib_insync==0.9.86

# Data Handling
pandas==2.2.0
numpy==1.26.4

# Data Sources
yfinance==0.2.38

# Visualization
matplotlib==3.8.3
plotly==5.20.0

# Technical Indicators
ta==0.11.0
# ta-lib==0.4.28  # Uncomment if you want TA-Lib (requires separate C library installation)

# Scientific Computing
scipy==1.12.0

# Utilities
python-dateutil==2.9.0
pytz==2024.1
requests==2.31.0

# Interactive Brokers Additional Tools
eventkit==1.0.3

# Async Support
nest-asyncio==1.6.0

# Configuration Management
python-dotenv==1.0.1

# Logging
loguru==0.7.2
